+++ admin/orders_status.php
@@ -132,7 +132,7 @@ if (!empty($action)) {
                                             FROM " . TABLE_ORDERS_STATUS . "
                                             WHERE language_id = " . (int)$_SESSION['languages_id'] . "
                                             ORDER BY sort_order ASC, orders_status_id ASC";
-                $orders_status_split = new splitPageResults($_GET['page'], zen_config('MAX_DISPLAY_SEARCH_RESULTS'), $orders_status_query_raw, $orders_status_query_numrows);
+                $orders_status_split = new splitPageResults($_GET['page'], (int)zen_config('MAX_DISPLAY_SEARCH_RESULTS'), $orders_status_query_raw, $orders_status_query_numrows);
                 $orders_status = $db->Execute($orders_status_query_raw);
                 foreach ($orders_status as $status) {
                   if ((!isset($_GET['oSID']) || (isset($_GET['oSID']) && ($_GET['oSID'] == $status['orders_status_id']))) && !isset($oInfo) && (substr($action, 0, 3) != 'new')) {
